fix routes; and move HandleNoSlashDir to component
